finish moving new parser from `parser` to `parse`
