+++ src/main.rs
@@ -13,7 +13,7 @@ async fn main() {
         ..Default::default()
     };
 
-    par::parser::set_miette_hook();
+    par::parse::set_miette_hook();
 
     eframe::run_native(
         "⅋layground",
