+++ src/par.rs
@@ -1,7 +1,6 @@
 pub mod language;
 pub mod lexer;
 pub mod parse;
-pub mod parser;
 pub mod process;
 pub mod runtime;
 pub mod types;
